Refactor: struct ParticipantsNum을 PokerGame으로 분리함             이유1 : 미션을 하면 할수록 ParticipantsNum에 대한 코드가 늘어남             따라서 ParticipantsNum에 대한 코드가 늘어나면             PokerGame.swift 파일이 더 읽기 어려워지므로 아예 분리함             이유2 : PokerGame을 통해 ParticipantsNum을 사용하는 것이 코드의             가독성을 해친다고 생각하기 때문이다.             before : PokerGame.ParticipantsNum             after : ParticipantsNum
